HL/KS/Shared-Backends: Move test utils to tests/utils.py
Move these functions only used in tests to shared.common.tests.utils:
 - create_finnish_social_security_number
 - normalize_whitespace
 - set_setting_to_value_or_del_with_none
 - utc_datetime

KS-Backend:
 - Move get_random_social_security_number_for_year to common.tests.utils

Refs YJDH-655 (Refactoring)
